+++ sample/src/androidMain/kotlin/com/github/panpf/zoomimage/sample/ui/gallery/PhotoPagerFragment.kt
@@ -133,7 +133,7 @@ class PhotoPagerFragment : BaseBindingFragment<FragmentPhotoPagerBinding>() {
         }
 
         binding.pager.apply {
-            offscreenPageLimit = ViewPager2.OFFSCREEN_PAGE_LIMIT_DEFAULT
+            offscreenPageLimit = 1
             appSettings.horizontalPagerLayout.collectWithLifecycle(viewLifecycleOwner) {
                 orientation =
                     if (it) ViewPager2.ORIENTATION_HORIZONTAL else ViewPager2.ORIENTATION_VERTICAL